Bind VariableFunctions as a module, not a class with static methods.
This was a bit trickier than I expected, because modules have
to be importable to be pickleable, but adding a module to another
module in the C API isn't really the right way to make it importable.
We hack around it by manually adding the module to sys.modules.

Thanks Richard Zou for an extremely useful prior attempt which helped
me make this work.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: c56258c579a47f4c07a2c551e0c81a30262acc90
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38136