Migrate OwnerRRef value store to generic torch Future
It's a followup of https://github.com/pytorch/pytorch/pull/32556, where an error handling boilerplate code path was added to the FutureMessage callback.

However, I noticed that the FutureMessage could never be set with an error, because the FutureMessage is a member in OwnerRRef,

- OwnerRRef does not have a setError method yet.
- The FutureMessage is only used for signaling
- The value of the RRef is contained in the `value_` field.

With the Future being generalized, it could contain more value types, not limited to Message.

This PR migrates the OwnerRRef value from the `value_` field to the generic Future.

In a later PR, it will be super easy to add a `setError` method for OwnerRRef, which calls `future_.setError(..)`. (I decide to do it later. I think it's better to migrate the call sites together with adding the new `setError` method.)

Also, this fixes the issue pointed out by https://github.com/pytorch/pytorch/pull/31086/files#r422256916.

This PR was submitted as https://github.com/pytorch/pytorch/pull/32608.

Differential Revision: [D5707692](https://our.internmc.facebook.com/intern/diff/D5707692/)

ghstack-source-id: 103757743
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38143